backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
pydantic==2.5.2
bcrypt==4.0.1
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.2
pymupdf==1.25.3
winsdk==1.0.0b10
